Got game_main to compile, calls game_loop() from QT
